refactor: applicants by id get request relocated
